+++ NumberProvider.hpp
@@ -7,16 +7,16 @@ namespace Lights
 	class NumberProvider : public ProviderBase
 	{
 	public:
-		inline NumberProvider( uint16_t value ) : providedValue( value ) {};
+		inline NumberProvider( int32_t value ) : providedValue( value ) {};
 
 		/// @brief Return the number held by this provider
 		///		   Although NumberProvider is fixed, it is the base class for other number providers
 		///		   that may support more complex behaviour.
 		///        Check for a self triggering.
 		/// @return 
-		inline virtual uint16_t Value()
+		inline virtual int32_t Value()
 		{
-			uint16_t retVal = providedValue;
+			int32_t retVal = providedValue;
 
 			// Is this is self triggering then get the next value
 			if ( selfIncrement == true )
@@ -29,13 +29,13 @@ namespace Lights
 
 		/// @brief Return the current value without going through Value()
 		/// @return 
-		inline virtual uint16_t GetValue() { return providedValue; }
+		inline virtual int32_t GetValue() { return providedValue; }
 
 		/// @brief Directly set the underlyinmg value
 		/// @param value 
-		inline virtual void SetValue( uint16_t value ) { providedValue = value; }
+		inline virtual void SetValue( int32_t value ) { providedValue = value; }
 
 	protected:
-		uint16_t providedValue = 0;
+		int32_t providedValue = 0;
 	};
 }
